feat : 폴더정리 및 코드정리
