--- app/_components/common/Header/fetchUser.tsx
@@ -1,9 +1,0 @@
-import API from "@/_api";
-import "@/_api";
-
-const fetchUser = async () => {
-	const response = await API["{teamId}/users/me"].GET({ teamId: "6-11" });
-	return response;
-};
-
-export default fetchUser;
